move WorkflowStatusInternal.Builder to test code
