+++ transact/src/test/java/dev/dbos/transact/database/SystemDatabaseTest.java
@@ -16,12 +16,12 @@ import dev.dbos.transact.exceptions.DBOSQueueDuplicatedException;
 import dev.dbos.transact.migrations.MigrationManager;
 import dev.dbos.transact.utils.DBUtils;
 import dev.dbos.transact.utils.PgContainer;
+import dev.dbos.transact.utils.WorkflowStatusInternalBuilder;
 import dev.dbos.transact.workflow.ForkOptions;
 import dev.dbos.transact.workflow.ScheduleStatus;
 import dev.dbos.transact.workflow.VersionInfo;
 import dev.dbos.transact.workflow.WorkflowSchedule;
 import dev.dbos.transact.workflow.WorkflowState;
-import dev.dbos.transact.workflow.internal.WorkflowStatusInternal;
 
 import java.time.Instant;
 import java.time.ZoneId;
@@ -56,7 +56,7 @@ public class SystemDatabaseTest {
   public void testDeleteWorkflows() throws Exception {
     for (var i = 0; i < 5; i++) {
       var wfid = "wfid-%d".formatted(i);
-      var status = WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING).build();
+      var status = WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING).build();
       sysdb.initWorkflowStatus(status, 5, false, false);
     }
 
@@ -139,7 +139,7 @@ public class SystemDatabaseTest {
   public void testDeleteWorkflowsList() throws Exception {
     for (var i = 0; i < 5; i++) {
       var wfid = "wfid-%d".formatted(i);
-      var status = WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING).build();
+      var status = WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING).build();
       sysdb.initWorkflowStatus(status, 5, false, false);
     }
 
@@ -156,15 +156,21 @@ public class SystemDatabaseTest {
     // Create workflows in different states
     for (var wfid : List.of("wf-pending-1", "wf-pending-2", "wf-pending-3")) {
       sysdb.initWorkflowStatus(
-          WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING).build(), 5, false, false);
+          WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING).build(),
+          5,
+          false,
+          false);
     }
     sysdb.initWorkflowStatus(
-        WorkflowStatusInternal.builder("wf-success", WorkflowState.PENDING).build(),
+        WorkflowStatusInternalBuilder.create("wf-success", WorkflowState.PENDING).build(),
         5,
         false,
         false);
     sysdb.initWorkflowStatus(
-        WorkflowStatusInternal.builder("wf-error", WorkflowState.PENDING).build(), 5, false, false);
+        WorkflowStatusInternalBuilder.create("wf-error", WorkflowState.PENDING).build(),
+        5,
+        false,
+        false);
     DBUtils.setWorkflowState(dataSource, "wf-success", WorkflowState.SUCCESS.name());
     DBUtils.setWorkflowState(dataSource, "wf-error", WorkflowState.ERROR.name());
 
@@ -194,16 +200,22 @@ public class SystemDatabaseTest {
     // Create workflows in different states
     for (var wfid : List.of("wf-cancelled-1", "wf-cancelled-2")) {
       sysdb.initWorkflowStatus(
-          WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING).build(), 5, false, false);
+          WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING).build(),
+          5,
+          false,
+          false);
       DBUtils.setWorkflowState(dataSource, wfid, WorkflowState.CANCELLED.name());
     }
     sysdb.initWorkflowStatus(
-        WorkflowStatusInternal.builder("wf-success", WorkflowState.PENDING).build(),
+        WorkflowStatusInternalBuilder.create("wf-success", WorkflowState.PENDING).build(),
         5,
         false,
         false);
     sysdb.initWorkflowStatus(
-        WorkflowStatusInternal.builder("wf-error", WorkflowState.PENDING).build(), 5, false, false);
+        WorkflowStatusInternalBuilder.create("wf-error", WorkflowState.PENDING).build(),
+        5,
+        false,
+        false);
     DBUtils.setWorkflowState(dataSource, "wf-success", WorkflowState.SUCCESS.name());
     DBUtils.setWorkflowState(dataSource, "wf-error", WorkflowState.ERROR.name());
 
@@ -275,7 +287,10 @@ public class SystemDatabaseTest {
   @Test
   public void testCancelWorkflowsNullInList() throws Exception {
     sysdb.initWorkflowStatus(
-        WorkflowStatusInternal.builder("wf-id", WorkflowState.PENDING).build(), 5, false, false);
+        WorkflowStatusInternalBuilder.create("wf-id", WorkflowState.PENDING).build(),
+        5,
+        false,
+        false);
 
     long beforeCancel = System.currentTimeMillis();
     sysdb.cancelWorkflows(Arrays.asList("wf-id", null));
@@ -288,7 +303,10 @@ public class SystemDatabaseTest {
   @Test
   public void testResumeWorkflowsNullInList() throws Exception {
     sysdb.initWorkflowStatus(
-        WorkflowStatusInternal.builder("wf-id", WorkflowState.PENDING).build(), 5, false, false);
+        WorkflowStatusInternalBuilder.create("wf-id", WorkflowState.PENDING).build(),
+        5,
+        false,
+        false);
     DBUtils.setWorkflowState(dataSource, "wf-id", WorkflowState.CANCELLED.name());
 
     long beforeResume = System.currentTimeMillis();
@@ -303,7 +321,10 @@ public class SystemDatabaseTest {
   @Test
   public void testDeleteWorkflowsNullInList() throws Exception {
     sysdb.initWorkflowStatus(
-        WorkflowStatusInternal.builder("wf-id", WorkflowState.PENDING).build(), 5, false, false);
+        WorkflowStatusInternalBuilder.create("wf-id", WorkflowState.PENDING).build(),
+        5,
+        false,
+        false);
 
     sysdb.deleteWorkflows(Arrays.asList("wf-id", null), false);
 
@@ -314,14 +335,14 @@ public class SystemDatabaseTest {
   public void testGetChildWorkflows() throws Exception {
     for (var i = 0; i < 5; i++) {
       var wfid = "wfid-%d".formatted(i);
-      var status = WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING).build();
+      var status = WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING).build();
       sysdb.initWorkflowStatus(status, 5, false, false);
     }
 
     for (var i = 0; i < 5; i++) {
       var parentWfId = "wfid-2";
       var wfid = "childwfid-%d".formatted(i);
-      var status = WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING).build();
+      var status = WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING).build();
       sysdb.initWorkflowStatus(status, 5, false, false);
       sysdb.recordChildWorkflow(
           parentWfId, wfid, i, "step-%d".formatted(i), System.currentTimeMillis());
@@ -330,7 +351,7 @@ public class SystemDatabaseTest {
     for (var i = 0; i < 5; i++) {
       var parentWfId = "childwfid-%d".formatted(i);
       var wfid = "grandchildwfid-%d".formatted(i);
-      var status = WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING).build();
+      var status = WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING).build();
       sysdb.initWorkflowStatus(status, 5, false, false);
       sysdb.recordChildWorkflow(
           parentWfId, wfid, i, "step-%d".formatted(i), System.currentTimeMillis());
@@ -351,7 +372,7 @@ public class SystemDatabaseTest {
   public void testRetries() throws Exception {
     var wfid = "wfid-1";
     var status =
-        WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING)
+        WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING)
             .workflowName("wf-name")
             .inputs("wf-inputs")
             .build();
@@ -381,7 +402,7 @@ public class SystemDatabaseTest {
   public void testDedupeId() throws Exception {
     var wfid = "wfid-1";
     var builder =
-        WorkflowStatusInternal.builder(wfid, WorkflowState.PENDING)
+        WorkflowStatusInternalBuilder.create(wfid, WorkflowState.PENDING)
             .workflowName("wf-name")
             .inputs("wf-inputs")
             .queueName("queue-name")
@@ -635,7 +656,7 @@ public class SystemDatabaseTest {
 
     // Create workflow status with authentication fields
     var status =
-        WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING)
+        WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING)
             .workflowName("TestWorkflow")
             .className("com.example.TestWorkflow")
             .authenticatedUser(authenticatedUser)
@@ -675,7 +696,7 @@ public class SystemDatabaseTest {
 
     // Create workflow status with null authentication fields
     var status =
-        WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING)
+        WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING)
             .workflowName("TestNullAuthWorkflow")
             .className("com.example.TestNullAuthWorkflow")
             .authenticatedUser(null)
@@ -710,7 +731,7 @@ public class SystemDatabaseTest {
 
     // Create workflow status with empty authenticated roles
     var status =
-        WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING)
+        WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING)
             .workflowName("TestEmptyRolesWorkflow")
             .className("com.example.TestEmptyRolesWorkflow")
             .authenticatedUser(authenticatedUser)
@@ -745,7 +766,7 @@ public class SystemDatabaseTest {
 
     // Create original workflow status with authentication fields
     var originalStatus =
-        WorkflowStatusInternal.builder(originalWorkflowId, WorkflowState.SUCCESS)
+        WorkflowStatusInternalBuilder.create(originalWorkflowId, WorkflowState.SUCCESS)
             .workflowName("OriginalTestWorkflow")
             .className("com.example.OriginalTestWorkflow")
             .authenticatedUser(authenticatedUser)
@@ -807,7 +828,7 @@ public class SystemDatabaseTest {
 
     // Create original workflow status with null authentication fields
     var originalStatus =
-        WorkflowStatusInternal.builder(originalWorkflowId, WorkflowState.SUCCESS)
+        WorkflowStatusInternalBuilder.create(originalWorkflowId, WorkflowState.SUCCESS)
             .workflowName("NullAuthTestWorkflow")
             .className("com.example.NullAuthTestWorkflow")
             .authenticatedUser(null)
@@ -855,7 +876,7 @@ public class SystemDatabaseTest {
 
     // Create original workflow status with empty authenticated roles
     var originalStatus =
-        WorkflowStatusInternal.builder(originalWorkflowId, WorkflowState.SUCCESS)
+        WorkflowStatusInternalBuilder.create(originalWorkflowId, WorkflowState.SUCCESS)
             .workflowName("EmptyAuthRolesTestWorkflow")
             .className("com.example.EmptyAuthRolesTestWorkflow")
             .authenticatedUser(authenticatedUser)
@@ -915,7 +936,7 @@ public class SystemDatabaseTest {
   @Test
   public void testWriteStreamAndReadStream() throws Exception {
     String workflowId = "stream-wf-1";
-    var status = WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING).build();
+    var status = WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING).build();
     sysdb.initWorkflowStatus(status, 5, false, false);
     int functionId = 1;
 
@@ -932,7 +953,7 @@ public class SystemDatabaseTest {
   @Test
   public void testWriteStreamFromWorkflow() throws Exception {
     String workflowId = "stream-wf-2";
-    var status = WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING).build();
+    var status = WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING).build();
     sysdb.initWorkflowStatus(status, 5, false, false);
     int functionId = 1;
 
@@ -945,7 +966,7 @@ public class SystemDatabaseTest {
   @Test
   public void testCloseStream() throws Exception {
     String workflowId = "stream-wf-3";
-    var status = WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING).build();
+    var status = WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING).build();
     sysdb.initWorkflowStatus(status, 5, false, false);
 
     sysdb.writeStreamFromWorkflow(workflowId, 1, "key1", "value1", "portable_json");
@@ -957,7 +978,7 @@ public class SystemDatabaseTest {
   @Test
   public void testGetAllStreamEntries() throws Exception {
     String workflowId = "stream-wf-4";
-    var status = WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING).build();
+    var status = WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING).build();
     sysdb.initWorkflowStatus(status, 5, false, false);
     int functionId = 1;
 
@@ -976,7 +997,7 @@ public class SystemDatabaseTest {
   @Test
   public void testReadStreamNotFound() throws Exception {
     String workflowId = "stream-wf-5";
-    var status = WorkflowStatusInternal.builder(workflowId, WorkflowState.PENDING).build();
+    var status = WorkflowStatusInternalBuilder.create(workflowId, WorkflowState.PENDING).build();
     sysdb.initWorkflowStatus(status, 5, false, false);
 
     assertThrows(IllegalArgumentException.class, () -> sysdb.readStream(workflowId, "key", 0));
@@ -987,7 +1008,7 @@ public class SystemDatabaseTest {
     assertThrows(
         NullPointerException.class,
         () -> {
-          var status = WorkflowStatusInternal.builder(null, WorkflowState.PENDING).build();
+          var status = WorkflowStatusInternalBuilder.create(null, WorkflowState.PENDING).build();
           sysdb.initWorkflowStatus(status, null, false, false);
         });
 
@@ -995,7 +1016,7 @@ public class SystemDatabaseTest {
     assertThrows(
         NullPointerException.class,
         () -> {
-          var status = WorkflowStatusInternal.builder("test-wf-id", null).build();
+          var status = WorkflowStatusInternalBuilder.create("test-wf-id", null).build();
           sysdb.initWorkflowStatus(status, null, false, false);
         });
   }
@@ -1007,7 +1028,7 @@ public class SystemDatabaseTest {
         IllegalStateException.class,
         () -> {
           var status =
-              WorkflowStatusInternal.builder("test-wf-1", WorkflowState.PENDING)
+              WorkflowStatusInternalBuilder.create("test-wf-1", WorkflowState.PENDING)
                   .workflowName("")
                   .build();
           sysdb.initWorkflowStatus(status, null, false, false);
@@ -1018,7 +1039,7 @@ public class SystemDatabaseTest {
         IllegalStateException.class,
         () -> {
           var status =
-              WorkflowStatusInternal.builder("test-wf-2", WorkflowState.PENDING)
+              WorkflowStatusInternalBuilder.create("test-wf-2", WorkflowState.PENDING)
                   .className("")
                   .build();
           sysdb.initWorkflowStatus(status, null, false, false);
@@ -1029,7 +1050,7 @@ public class SystemDatabaseTest {
         IllegalStateException.class,
         () -> {
           var status =
-              WorkflowStatusInternal.builder("test-wf-3", WorkflowState.PENDING)
+              WorkflowStatusInternalBuilder.create("test-wf-3", WorkflowState.PENDING)
                   .instanceName("")
                   .build();
           sysdb.initWorkflowStatus(status, null, false, false);
@@ -1040,7 +1061,7 @@ public class SystemDatabaseTest {
         IllegalStateException.class,
         () -> {
           var status =
-              WorkflowStatusInternal.builder("test-wf-4", WorkflowState.PENDING)
+              WorkflowStatusInternalBuilder.create("test-wf-4", WorkflowState.PENDING)
                   .queueName("")
                   .build();
           sysdb.initWorkflowStatus(status, null, false, false);
@@ -1051,7 +1072,7 @@ public class SystemDatabaseTest {
         IllegalStateException.class,
         () -> {
           var status =
-              WorkflowStatusInternal.builder("test-wf-5", WorkflowState.PENDING)
+              WorkflowStatusInternalBuilder.create("test-wf-5", WorkflowState.PENDING)
                   .deduplicationId("")
                   .build();
           sysdb.initWorkflowStatus(status, null, false, false);
@@ -1062,7 +1083,7 @@ public class SystemDatabaseTest {
         IllegalStateException.class,
         () -> {
           var status =
-              WorkflowStatusInternal.builder("test-wf-6", WorkflowState.PENDING)
+              WorkflowStatusInternalBuilder.create("test-wf-6", WorkflowState.PENDING)
                   .queuePartitionKey("")
                   .build();
           sysdb.initWorkflowStatus(status, null, false, false);
@@ -1073,7 +1094,7 @@ public class SystemDatabaseTest {
   public void testInsertWorkflowStatusValidNullValues() throws Exception {
     // Test that null values (except for required fields) are allowed
     var status =
-        WorkflowStatusInternal.builder("test-valid-nulls", WorkflowState.PENDING)
+        WorkflowStatusInternalBuilder.create("test-valid-nulls", WorkflowState.PENDING)
             .workflowName(null)
             .className(null)
             .instanceName(null)
@@ -1092,7 +1113,7 @@ public class SystemDatabaseTest {
   public void testInsertWorkflowStatusValidNonEmptyValues() throws Exception {
     // Test that non-empty values are allowed
     var status =
-        WorkflowStatusInternal.builder("test-valid-values", WorkflowState.PENDING)
+        WorkflowStatusInternalBuilder.create("test-valid-values", WorkflowState.PENDING)
             .workflowName("TestWorkflow")
             .className("com.example.TestWorkflow")
             .instanceName("test-instance")
